tests/test_langchain_agent.py
import pytest
from src.langchain_agent import LangChainAgent

def test_langchain_agent_process_java_code():
    agent = LangChainAgent()
    java_code = """
    public class Example {
        public void oldMethod() {
            System.out.println("Hello, World!");
        }
    }
    """
    suggestions = agent.process_java_code(java_code)
    
    assert isinstance(suggestions, list)
    assert len(suggestions) > 0
    for suggestion in suggestions:
        assert 'file_path' in suggestion
        assert 'description' in suggestion
        assert 'start_line' in suggestion
        assert 'end_line' in suggestion
        assert 'original_snippet' in suggestion
        assert 'modified_code' in suggestion
        assert 'difficulty_level' in suggestion

def test_langchain_agent_empty_code():
    agent = LangChainAgent()
    java_code = ""
    suggestions = agent.process_java_code(java_code)
    
    assert suggestions == []  # Expecting no suggestions for empty code

def test_langchain_agent_invalid_code():
    agent = LangChainAgent()
    java_code = "invalid java code"
    suggestions = agent.process_java_code(java_code)
    
    assert isinstance(suggestions, list)
    assert len(suggestions) == 0  # Expecting no suggestions for invalid code