requirements.txt
langchain
pydantic
python-dotenv
langchain-google-genai

